Bundled anemic structs into a types.go file to save on vertical space
