--- cancel.go
@@ -1,20 +1,0 @@
-package adyen
-
-// Cancel structure for Cancel request
-type Cancel struct {
-	Reference         string `json:"reference"`
-	MerchantAccount   string `json:"merchantAccount"`
-	OriginalReference string `json:"originalReference"`
-}
-
-// CancelResponse is a response structure for Adyen cancellation
-type CancelResponse struct {
-	PspReference string `json:"pspReference"`
-	Response     string `json:"response"`
-}
-
-// CancelOrRefundResponse is a response structure for Adyen cancelOrRefund
-type CancelOrRefundResponse struct {
-	PspReference string `json:"pspReference"`
-	Response     string `json:"response"`
-}
