--- capture.go
@@ -1,15 +1,0 @@
-package adyen
-
-// Capture structure for Capture request
-type Capture struct {
-	ModificationAmount *Amount `json:"modificationAmount"`
-	Reference          string  `json:"reference"`
-	MerchantAccount    string  `json:"merchantAccount"`
-	OriginalReference  string  `json:"originalReference"`
-}
-
-// CaptureResponse is a response structure for Adyen capture
-type CaptureResponse struct {
-	PspReference string `json:"pspReference"`
-	Response     string `json:"response"`
-}
